basic features

+++ routes/api.php
@@ -1,7 +1,10 @@
 <?php
 
 use App\Http\Controllers\Api\v1\AuthController;
-use App\Http\Controllers\PostController;
+use App\Http\Controllers\api\v1\UserController;
+use App\Http\Controllers\v1\CategoryController;
+use App\Http\Controllers\v1\CommentController;
+use App\Http\Controllers\v1\PostController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
@@ -20,14 +23,41 @@ Route::group([
     Route::post('/profile', [AuthController::class, 'profile'])->middleware('auth:api');
 });
 
-Route::get('/posts',[PostController::class,'index']);
-Route::post('/posts',[PostController::class,'store']);
-Route::get('/posts/{post}',[PostController::class,'show']);
-Route::put('/posts/{post}',[PostController::class,'update']);
-Route::delete('/posts/{post}',[PostController::class,'destroy']);
+Route::group(['prefix' => '/v1'],function(){
 
-Route::get('/posts/{post}/comments',[PostController::class,'index']);
-Route::post('/posts/{post}/comments',[PostController::class,'store']);
-Route::get('/posts/{post}/comments/{comment}',[PostController::class,'show']);
-Route::put('/posts/{post}/comments/{comment}',[PostController::class,'update']);
-Route::delete('/posts/{post}/comments/{comment}',[PostController::class,'destroy']);
+    Route::get('/posts',[PostController::class,'index']);
+    Route::get('/posts/{post}',[PostController::class,'show']);
+    Route::post('/posts',[PostController::class,'store'])->middleware('auth:api');
+    Route::put('/posts/{post}',[PostController::class,'update'])->middleware('auth:api')->can('update_post','post');
+    Route::delete('/posts/{post}',[PostController::class,'destroy'])->middleware('auth:api')->can('update_post','post');
+
+    Route::get('/posts/{post}/comments',[CommentController::class,'index']);
+    Route::get('/posts/{post}/comments/{comment}',[CommentController::class,'show']);
+    Route::post('/posts/{post}/comments',[CommentController::class,'store'])->middleware('auth:api');
+    Route::put('/posts/{post}/comments/{comment}',[CommentController::class,'update'])->middleware('auth:api')->can('update_comment','comment');
+    Route::delete('/posts/{post}/comments/{comment}',[CommentController::class,'destroy'])->middleware('auth:api')->can('delete_comment','comment');
+
+
+    Route::get('/categories', [CategoryController::class, 'index']);      
+    Route::get('/categories/{category}', [CategoryController::class, 'show']); //lists all posts for a speicific category
+    Route::group([
+        'prefix' => 'admin',
+        'middleware' => ['auth:api', 'can:admin_only'] 
+    ], function () {
+
+        Route::post('categories', [CategoryController::class, 'store']);
+        Route::put('categories/{category}', [CategoryController::class, 'update']);
+        Route::delete('categories/{category}', [CategoryController::class, 'destroy']);   
+
+        Route::post('users', [UserController::class, 'store']);
+        Route::put('users/{user}', [UserController::class, 'update']);
+        Route::delete('users/{user}', [UserController::class, 'destroy']);
+    });
+
+    Route::get('/admin/users', [UserController::class, 'index']);      
+    Route::get('/admin/users/{user}', [UserController::class, 'show']);
+
+});
+
+
+
